+++ examples/plot_zt.py
@@ -6,6 +6,9 @@ zt-plots.
 This example shows how to create a zt-plot of a given DU and event ID.
 """
 
+# Author: Tamas Gal <tgal@km3net.de>
+# License: BSD-3
+
 import pandas as pd
 import matplotlib.pyplot as plt
 import km3pipe as kp
@@ -24,14 +27,10 @@ geometry.apply(hits)
 
 fig, ax = plt.subplots()
 
-hits[hits['du'] == DU].plot('time', 'pos_z', style='.', ax=ax)
+hits[hits['du'] == DU].plot('time', 'pos_z', style='.', ax=ax, label='hits')
 triggered_hits = hits[(hits['du'] == DU) & (hits['triggered'] == True)]
-triggered_hits.plot('time', 'pos_z', style='.', ax=ax)
+triggered_hits.plot('time', 'pos_z', style='.', ax=ax, label='triggered hits')
 
-try:
-    ax.legend_.remove()
-except AttributeError:
-    pass
 ax.set_title("zt-plot of event {0} on DU{1}".format(EVENT_ID, DU))
 ax.set_xlabel("time [ns]")
 ax.set_ylabel("z [m]")